replacing with int, where months

+++ src/main/java/ru/Netology/javaqa/ArrayOfMonthlySales/services/StatsService.java
@@ -6,7 +6,7 @@ public class StatsService {
         //сумма всех продаж
         long total = 0;
         for (long i = 0; i < sales.length; i++) {
-            total += sales[Math.toIntExact((long) i)];
+            total += sales[Math.toIntExact(i)];
         }
         return total;
     }
@@ -15,17 +15,17 @@ public class StatsService {
         //средняя сумма продаж в месяц
         long total = 0;
         for (long i = 0; i < sales.length; i++) {
-            total += sales[Math.toIntExact((long) i)];
+            total += sales[Math.toIntExact(i)];
         }
         return (total / sales.length);
     }
 
     public long minSales(long[] sales) {
         //минимум продаж
-        long minMonth = 0;
+        int minMonth = 0;
         for (long i = 0; i < sales.length; i++) {
-            if (sales[Math.toIntExact((long) i)] <= sales[Math.toIntExact((long) minMonth)]) {
-                minMonth = i;
+            if (sales[Math.toIntExact(i)] <= sales[Math.toIntExact(minMonth)]) {
+                minMonth = (int) i;
             }
         }
         return (minMonth + 1);
@@ -33,10 +33,10 @@ public class StatsService {
 
     public long maxSales(long[] sales) {
         // максимум продаж
-        long maxMonth = 0;
+        int maxMonth = 0;
         for (long i = 0; i < sales.length; i++) {
-            if (sales[Math.toIntExact((long) i)] >= sales[Math.toIntExact((long) maxMonth)]) {
-                maxMonth = i;
+            if (sales[Math.toIntExact(i)] >= sales[Math.toIntExact(maxMonth)]) {
+                maxMonth = (int) i;
             }
         }
         return (maxMonth + 1);
@@ -44,10 +44,10 @@ public class StatsService {
 
     public long belowAverage(long[] sales) {
         //продажи ниже среднего
-        long numberOfMonth = 0;
-        long average = averageSalesAmount (sales);
+        int numberOfMonth = 0;
+        long average = averageSalesAmount(sales);
         for (long i = 0; i < sales.length; i++) {
-            if (sales[Math.toIntExact((long) i)] >= average) {
+            if (sales[Math.toIntExact(i)] >= average) {
                 numberOfMonth = numberOfMonth + 1;
             }
         }
@@ -56,10 +56,10 @@ public class StatsService {
 
     public long aboveAverage(long[] sales) {
         //продажи выше среднего
-        long numberOfMonth = 0;
+        int numberOfMonth = 0;
         long average = averageSalesAmount(sales);
         for (long i = 0; i < sales.length; i++) {
-            if (sales[Math.toIntExact((long) i)] <= average) {
+            if (sales[Math.toIntExact(i)] <= average) {
                 numberOfMonth = numberOfMonth + 1;
             }
         }
